+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.2.1 (2016-09-24)
+
+- jonpeterson: Fixed merging of field and accessor annotations.
+- jonpeterson: Added a property to @JsonVersionedModel to specify a model version in which serialization of the modelVersion should be supressed.
+
 ## 1.2.0 (2016-09-22)
 
 - brentryan:   Made @JsonVersionedModel.toCurrentConverter optional for when used on version 1 of a model.